# perf(es/minifier): Optimize precompress optimizer