--- a/crates/swc_ecma_minifier/src/pass/precompress.rs
+++ b/crates/swc_ecma_minifier/src/pass/precompress.rs
@@ -77,17 +77,14 @@ impl PrecompressOptimizer<'_> {
             return;
         }
 
-        maybe_par!(
-            stmts.iter_mut().for_each(|stmt| {
-                stmt.visit_mut_with(&mut PrecompressOptimizer {
-                    options: self.options,
-                    module_info: self.module_info,
-                    marks: self.marks,
-                    ctx: self.ctx,
-                })
-            }),
-            *crate::HEAVY_TASK_PARALLELS
-        );
+        stmts.iter_mut().for_each(|stmt| {
+            stmt.visit_mut_with(&mut PrecompressOptimizer {
+                options: self.options,
+                module_info: self.module_info,
+                marks: self.marks,
+                ctx: self.ctx,
+            })
+        })
     }
 }